+++ narrenschiff/narrenschiff.py
@@ -1,6 +1,7 @@
 import click
 
 import narrenschiff
+from narrenschiff.cli.env import env
 from narrenschiff.cli.sail import sail
 from narrenschiff.cli.chest import chest
 from narrenschiff.cli.secretmap import secretmap
@@ -26,6 +27,7 @@ def narrenschiff(verbosity=0):
     logger.set_verbosity(verbosity)
 
 
+narrenschiff.add_command(env)
 narrenschiff.add_command(sail)
 narrenschiff.add_command(chest)
 narrenschiff.add_command(secretmap)