+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
 * You can now check version with `--version` flag
+* You can print environment info in case you need this for debugging or issue reporting, this option includes output optionally formatted with Markdown
 
 # 1.3.0 - 17.6.2020.